From 5bb04a60d19ddf150335e29285d19bcbe87d343d Mon Sep 17 00:00:00 2001 From: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com> Date: Fri, 20 May 2011 12:02:14 -0300 Subject: [PATCH 1/2] Staging: vt6656: avoid crashing with wrong channel information
Current vt6656 code in kernel doesn't set and update byCurrentCh, so
default initialization keeps it zeroed. One side effect of this is that
it can lead to an oops, like the one reported at
http://bugs.launchpad.net/bugs/768524

In this specific bug, after analysing the crash, the oops happens when
we call RFbSetPower(pDevice, wCurrentRate, pDevice->byCurrentCh).
Because byCurrentCh is zero, we get a crash in
"byPwr = pDevice->abyCCKPwrTbl[uCH-1]" assignment, as uCh is 0.

This patch thus properly set/update byCurrentCh. A newer version of
vt6656 driver released by Via (v1.20.03) [1] has a similar change.

[1] http://www.viaarena.com/Driver/VT6656_linux_src_v1.20.03_x86.rar

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>